feat(relationship): 점수 상한을 제거한다 (#28)

+++ apps/ios/Woorisai/App/WoorisaiColor.swift
@@ -273,7 +273,7 @@ enum WoorisaiColor {
     case decrease
     case unchanged
 
-    init(_ value: Int) {
+    init(_ value: Int64) {
       if value > 0 {
         self = .increase
       } else if value < 0 {
